Abstract our request handler resolution
Currently the default is to inspect all the assemblies in the current appdomain for types implementing `IRequestHandler` but it would be nice to be able to customize this resolution.  This abstracts the concept out so you can provide your own resolver and suggest the types explicitly that should be used.